+++ packages/raga-app/src/client/components/common/tree.module.scss
@@ -1,7 +1,15 @@
 @use "@blueprintjs/core/lib/scss/variables.scss" as bp;
 
 .node {
-  // empty
+  &.selectedPath {
+    > :global(.#{bp.$ns}-tree-node-content) {
+      background-color: darken(bp.$pt-intent-primary, 15%);
+    }
+
+    &:global(.#{bp.$ns}-tree-node-selected) > :global(.#{bp.$ns}-tree-node-content) {
+      background-color: bp.$pt-intent-primary;
+    }
+  }
 }
 
 :global(.#{bp.$ns}-dark) .node {
