+++ packages/raga-app/src/client/views/libraryView.tsx
@@ -1,8 +1,6 @@
 import { Card, NonIdealState, ProgressBar, Text } from "@blueprintjs/core";
 import classNames from "classnames";
-import { useRef, useState } from "react";
 import { Panel, PanelGroup } from "react-resizable-panels";
-import { useEffectOnce } from "usehooks-ts";
 
 import { formatStatNumber } from "../../common/format";
 import { AudioPlayer } from "../components/audioPlayer/audioPlayer";
@@ -54,19 +52,11 @@ export default function LibraryView() {
 }
 
 function Library() {
-  const [headerHeight, setHeaderHeight] = useState<number>(0);
-  const headerRef = useRef<HTMLDivElement>(null);
   const selectedPlaylistId = appStore.use.selectedPlaylistId();
-
-  useEffectOnce(() => {
-    if (headerRef.current != null) {
-      setHeaderHeight(headerRef.current.clientHeight);
-    }
-  });
 
   return (
     <div className={classNames("flex-column", styles.library)}>
-      <div className={styles.libraryHeader} ref={headerRef}>
+      <div className={styles.libraryHeader}>
         <LibraryHeaderSection className={styles.libraryHeaderSection} />
       </div>
       <div className={styles.audioPlayer}>
@@ -74,7 +64,7 @@ function Library() {
       </div>
       <PanelGroup direction="horizontal">
         <Panel className={styles.librarySidebar} defaultSize={20} minSize={20} maxSize={40}>
-          <PlaylistTable headerHeight={headerHeight} />
+          <PlaylistTable />
           <LibrarySidebarFooter />
         </Panel>
         <ResizeHandle />
